core: convert json-payload to typescript
