+++ packages/core/package.json
@@ -10,7 +10,6 @@
       "default": "./dist/index.cjs"
     },
     "./lib/clone-client": "./src/lib/clone-client.js",
-    "./lib/json-payload": "./src/lib/json-payload.js",
     "./lib/es-utils/assign": "./src/lib/es-utils/assign.js",
     "./lib/validators/int-range": "./src/lib/validators/int-range.js",
     "./lib/es-utils/map": "./src/lib/es-utils/map.js",
