+++ packages/core/src/index.ts
@@ -3,3 +3,6 @@ export { default as Client } from './client'
 export { default as Event } from './event'
 export { default as Session } from './session'
 export { default as schema } from './config'
+export { default as jsonPayload } from './lib/json-payload'
+
+export * from './common'
