--- packages/core/src/lib/json-payload.d.ts
@@ -1,10 +1,0 @@
-type RedactedKey = string | RegExp
-
-interface JsonPayload {
-  event: (event: Object, redactedKeys?: RedactedKey[]) => string
-  session: (session: Object, redactedKeys?: RedactedKey[]) => string
-}
-
-declare const jsonPayload: JsonPayload
-
-export default jsonPayload
