+++ packages/core/src/safe-json-stringify.d.ts
@@ -1,0 +1,1 @@
+declare module "@bugsnag/safe-json-stringify";
