+++ packages/core/test/lib/json-payload.test.ts
@@ -51,7 +51,7 @@ describe('jsonPayload.event', () => {
 
     payload.events[0]._metadata = { 'big thing': makeBigObject() }
 
-    expect(jsonPayload.event(payload)).toBe('{"api_key":"d145b8e5afb56516423bc4d605e45442","events":[{"errorMessage":"Failed load tickets","errorClass":"CheckoutError","user":{"name":"Jim Bug","email":"jim@bugsnag.com"},"request":{"api_key":"245b39ebd3cd3992e85bffc81c045924"},"_metadata":{"notifier":"WARNING!\\nSerialized payload was 2.003435MB (limit = 1MB)\\nmetadata was removed"}}]}')
+    expect(jsonPayload.event(payload)).toBe('{"api_key":"d145b8e5afb56516423bc4d605e45442","events":[{"errorMessage":"Failed load tickets","errorClass":"CheckoutError","user":{"name":"Jim Bug","email":"jim@bugsnag.com"},"request":{"api_key":"245b39ebd3cd3992e85bffc81c045924"},"_metadata":{"notifier":"WARNING!\\n  Serialized payload was 2.003435MB (limit = 1MB)\\n  metadata was removed"}}]}')
   })
 
   it('does not attempt to strip any other data paths from the payload to reduce the size', () => {
