+++ packages/core/tsconfig.json
@@ -9,5 +9,5 @@
     "emitDeclarationOnly": true,
     "declarationDir": "dist/types",
   },
-  "include": ["src/**/*.ts", "src/**/*.js"]
+  "include": ["src/**/*.ts", "src/**/*.js", "safe-json-stringify.d.ts"]
 }
