+++ packages/delivery-x-domain-request/rollup.config.npm.mjs
@@ -2,5 +2,5 @@ import createRollupConfig from "../../.rollup/index.mjs";
 
 export default createRollupConfig({
   input: "src/delivery.ts",
-  external: ['@bugsnag/core', '@bugsnag/core/lib/json-payload']
+  external: ['@bugsnag/core']
 });
